# guit-app/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Guit App API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v2/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered and token pair issued" },
          "400": { "description": "Validation error" },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/api/v2/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login success and token pair issued" },
          "400": { "description": "Validation error" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/v2/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Issue a new access/refresh token pair",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": { "type": "string" }
                },
                "required": ["refreshToken"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token pair refreshed" },
          "400": { "description": "Validation error" },
          "401": { "description": "Invalid/expired token" }
        }
      }
    },
...
  }
}